+++ src/component/index.ts
@@ -1,5 +1,7 @@
 export * from "./Anchor";
 export * from "./ErrorPage";
+export * from "./IconInput";
 export * from "./Link";
 export * from "./Login";
 export * from "./LoginWrapper";
+export * from "./TablerIcon";