+++ src/util/index.ts
@@ -3,7 +3,4 @@ export * from "./auth";
 export * from "./interface";
 export * from "./session";
 export * from "./theme";
-export * from "./leadingSlash";
-export * from "./safeJSON";
-export * from "./validateForm";
-export * from "./variableName";
+export * from "./validateFormThrow";