+++ src/util/interface/index.ts
@@ -1,2 +1,3 @@
 export * from "./api/APIProp";
+export * from "./api/APIRoute";
 export * from "./route/RouteOptions";